Restructuring analysis project (#165)
* updating analysis environment and beginning updates to project

* beginning analysis fixes

* restructuring analysis project

* updating pyproject typeo args to match new structure

* getting analysis working, fixes issues in other projects

* increasing parallelization on analysis

* getting new analysis working

* appending parameters to injection timeslides rather than overwriting

* centering injected waveforms

* initial commit of vizapp

* getting initial parts of app working

* adding tap tool and callback to distribution plot

* basic vizapp functionality working

* add binning for snr x-axis

* fix box select bug

* add distance vs efficiency; fix bug in x range

* get event inspector working

* run pre-commit on vizapp

* fix event inspector

* getting app running with args

* taking waveform roll out of training project

* fixing tests

* getting injection tests passing

* fixing glitch insertion issue

* quick update to glitch insert bug

+++ projects/sandbox/export/export.py
@@ -1,4 +1,5 @@
 import logging
+import os
 import shutil
 from pathlib import Path
 from typing import Callable, Optional
@@ -123,7 +124,7 @@ def export(
     # instantiate a model repository at the
     # indicated location and see if a bbhnet
     # model already exists in this repository
-    if clean:
+    if clean and os.path.exists(repository_directory):
         shutil.rmtree(repository_directory)
     repo = qv.ModelRepository(repository_directory)
     try:
